api/requirements.txt
fastapi==0.95.0
fastapi-versioning==0.10.0
gunicorn==20.1.0
influxdb==5.3.1
influxdb-client==1.34.0
PyMySQL==1.0.3
mysqlclient==2.1.1
pandas==1.5.1
requests==2.28.1
swagger-ui==0.1.2
uvicorn==0.21.1
python-dotenv==0.19.0
